Computer Vision/requirements.txt
opencv-python
picamera2
ultralytics
numpy
scikit-image
RPi.GPIO
